ArchCompStochasticModels

Documentation for ArchCompStochasticModels.

ArchCompStochasticModels.DiscreteTimeStochasticHybridSystem

ArchCompStochasticModels.DiscreteTimeStochasticSystem

ArchCompStochasticModels.automated_anaesthesia

ArchCompStochasticModels.controlled_gaussian_van_der_pol

ArchCompStochasticModels.controlled_van_der_pol

ArchCompStochasticModels.cs1_bas

ArchCompStochasticModels.cs2_bas

ArchCompStochasticModels.fully_automated_anaesthesia

ArchCompStochasticModels.integrator_chain

ArchCompStochasticModels.van_der_pol

ArchCompStochasticModels.DiscreteTimeStochasticHybridSystem — Type

DiscreteTimeStochasticHybridSystem

A discrete-time stochastic hybrid system in the style of [1] is a tuple $H = (\mathcal{Q}, n, \mathcal{U}, \Sigma, T_x, T_q, R)$ where:

$\mathcal{Q} = \{q_1, q_2, \ldots, q_m\}$ is a finite set of discrete modes;

$n : \mathcal{Q} \to \mathbb{N}$ is the number of continuous dimensions in each mode. The hybrid state space is defined as $\mathcal{S} = \bigcup_{q \in \mathcal{Q}} \{q\} \times \mathbb{R}^{n(q)}$;

$\mathcal{U}$ is a compact Borel space representing the transition control space;

$\Sigma$ is a compact Borel space representing the reset control space;

$T_x : \mathcal{B}(\mathbb{R}^{n(\cdot)}) \times \mathcal{S} \times \mathcal{U} \to [0, 1]$ is the continuous transition kernel;

$T_q : \mathcal{Q} \times \mathcal{S} \times \mathcal{U} \to [0, 1]$ is the discrete transition kernel, and

$R : \mathcal{B}(\mathbb{R}^{n(\cdot)}) \times \mathcal{S} \times \Sigma \times \mathcal{Q} \to [0, 1]$ is the reset transition kernel.

Given a Markov policy $\mu : \mathcal{S} \to \mathcal{U} \times \Sigma$, the system that evolves according to the following at each time step $k$ with state $s_k = (q_k, x_k)$:

Let $u_k, \sigma_k = \mu(s_k)$;

Extract the (potential) mode transition $q_{k+1} \sim T_q(\cdot \mid s_k, u_k)$;

If $q_{k+1} = q_k$, then extract the continuous state evolution as $x_{k + 1} \sim T_x(\cdot \mid s_k, u_k)$, and

If $q_{k+1} \neq q_k$, then extract the continuous state evolution as $x_{k + 1} \sim R(\cdot \mid s_k, \sigma_k, q_{k+1})$.

Warning

The transition and reset control space are assumed to be compact but it is not enforced in the type definition.

Fields

Example

[1] Abate, A., Prandini, M., Lygeros, J., & Sastry, S. (2008). Probabilistic reachability and safety for controlled discrete time stochastic hybrid systems. Automatica, 44(11), 2724-2734.

ArchCompStochasticModels.DiscreteTimeStochasticSystem — Type

DiscreteTimeStochasticSystem

A discrete-time stochastic system is a tuple $H = (\mathcal{X}, \mathcal{U}, T_x)$ where:

$\mathcal{X} \subseteq \mathbb{R}^n$ is the state space;

$\mathcal{U}$ is a compact Borel space representing the control space, and

$T_x : \mathcal{B}(\mathbb{R}^{n(\cdot)}) \times \mathcal{X} \times \mathcal{U} \to [0, 1]$ is the continuous transition kernel.

Given a Markov policy $\mu : \mathcal{X} \to \mathcal{U}$, the system that evolves according to the following at each time step $k$:

Let $u_k = \mu(x_k)$, and

Extract the continuous state evolution as $x_{k + 1} \sim T_x(\cdot \mid x_k, u_k)$.

Warning

The control space are assumed to be compact but it is not enforced in the type definition.

Fields

Example

ArchCompStochasticModels.automated_anaesthesia — Method

Automated Anaesthesia Delievery System Benchmark

The concentration of Propofol in different compartments of the body are modelled using the three-compartment pharmacokinetic system. An automated system controls the basal dosage of Propofol to the patient, while an aneasthesiologist can administer a bolus dose. The system model includes a stochastic model of the aneasthesiologist behavior based on the concentration and the number of bolus doses administered. The hybrid system behavior arises from this aneasthesiologist behavior.

First presented in Abate, A., Blom, H., Cauchi, N., Hartmanns, A., Lesser, K., Oishi, M., ... & Vinod, A. P. (2018). ARCH-COMP18 category report: Stochastic modelling. In 5th International Workshop on Applied Verification of Continuous and Hybrid Systems, ARCH 2018 (pp. 71-103). EasyChair.

Mathematical Model

\[\begin{aligned} \bar{x}[k + 1] &= \begin{bmatrix} 0.8192 & 0.03412 & 0.01265 \\ 0.01646 & 0.9822 & 0.0001 \\ 0.0009 & 0.00002 & 0.9989 \end{bmatrix} \bar{x}[k] + \begin{bmatrix} 0.01883 \\ 0.0002 \\ 0.00001 \end{bmatrix} (v[k] + \sigma[k]) + w[k] \\ &= A \bar{x}[k] + B (v[k] + \sigma[k]) + w[k] \end{aligned}\]

where $\bar{x}[k]$ is the continuous state vector, $v[k]$ is the automated delivery system control input, $w[k] \sim \mathcal{N}(0, M)$ is the process noise, and $\sigma[k]$ is the noise from the aneasthesiologist. $\sigma[k]$ is a semi-Markov random variable taking values in $\{0, 30\}$ that represents the aneasthesiologist's decision to administer a bolus dose. To make the stochastic system Markovian, a binary state vector $q[k]$ is introduced that represents the aneasthesiologist's decision to administer a bolus dose at time $k$. Then the stochastic decision to administer a bolus dose is conditioned on $\bar{x}_1[k]$ and $q[k]$.

ArchCompStochasticModels.controlled_gaussian_van_der_pol — Method

Controlled Gaussian Van der Pol Oscillator Benchmark

First presented in Alessandro, A., Henk, B., Nathalie, C., Joanna, D., Arnd, H., Mahmoud, K., ... & Zuliani, P. (2020). ARCH-COMP20 Category Report: Stochastic Models.

Mathematical Model

\[\begin{aligned} x_1[k + 1] &= x_1[k] + \tau x_2[k] + w[k]\\ x_2[k + 1] &= x_2[k] + \tau ((1 - x_1[k]^2) x_2[k] - x_1[k]) + u[k] w[k] \end{aligned}\]

ArchCompStochasticModels.controlled_van_der_pol — Method

Controlled Stochastic Van der Pol Oscillator Benchmark

First presented in Alessandro, A., Henk, B., Nathalie, C., Joanna, D., Arnd, H., Mahmoud, K., ... & Zuliani, P. (2020). ARCH-COMP20 Category Report: Stochastic Models.

Mathematical Model

\[\begin{aligned} x_1[k + 1] &= x_1[k] + \tau x_2[k] + w[k]\\ x_2[k + 1] &= x_2[k] + \tau ((1 - x_1[k]^2) x_2[k] - x_1[k]) + u[k] w[k] \end{aligned}\]

ArchCompStochasticModels.cs1_bas — Method

Building Automation System 2D Benchmark

A building automation system (BAS) with two zones, each heated by one radiator and with a shared air supply.

First presented in Abate, A., Blom, H., Cauchi, N., Hartmanns, A., Lesser, K., Oishi, M., ... & Vinod, A. P. (2018). ARCH-COMP19 category report: Stochastic modelling. In 5th International Workshop on Applied Verification of Continuous and Hybrid Systems, ARCH 2018 (pp. 71-103). EasyChair. Concrete values taken from Abate, A., Blom, H., Cauchi, N., Degiorgio, K., Fraenzle, M., Hahn, E. M., ... & Vinod, A. P. (2019). ARCH-COMP19 category report: Stochastic modelling. In 6th International Workshop on Applied Verification of Continuous and Hybrid Systems, ARCH 2019 (pp. 62-102). EasyChair.

Mathematical Model

\[\begin{aligned} x[k + 1] &= A x[k] + B u[k] + Q + B_w w[k]\\ y[k] &= \begin{bmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \end{bmatrix} x[k] \end{aligned}\]

ArchCompStochasticModels.cs2_bas — Method

Building Automation System 7D Benchmark

A building automation system (BAS).

First presented in Abate, A., Blom, H., Cauchi, N., Hartmanns, A., Lesser, K., Oishi, M., ... & Vinod, A. P. (2018). ARCH-COMP19 category report: Stochastic modelling. In 5th International Workshop on Applied Verification of Continuous and Hybrid Systems, ARCH 2018 (pp. 71-103). EasyChair.

\[\begin{aligned} x[k + 1] &= A x[k] + B u[k] + Q + B_w w[k]\\ y[k] &= \begin{bmatrix} 1 & 0 & 0 & 0 & 0 & 0 & 0 \end{bmatrix} x[k] \end{aligned}\]

ArchCompStochasticModels.fully_automated_anaesthesia — Method

Fully-Automated Anaesthesia Delievery System Benchmark

The concentration of Propofol in different compartments of the body are modelled using the three-compartment pharmacokinetic system. An fully-automated system controls the dosage of Propofol to the patient.

First presented in Abate, A., Blom, H., Cauchi, N., Hartmanns, A., Lesser, K., Oishi, M., ... & Vinod, A. P. (2018). ARCH-COMP18 category report: Stochastic modelling. In 5th International Workshop on Applied Verification of Continuous and Hybrid Systems, ARCH 2018 (pp. 71-103). EasyChair.

Mathematical Model

\[\begin{aligned} \bar{x}[k + 1] &= \begin{bmatrix} 0.8192 & 0.03412 & 0.01265 \\ 0.01646 & 0.9822 & 0.0001 \\ 0.0009 & 0.00002 & 0.9989 \end{bmatrix} \bar{x}[k] + \begin{bmatrix} 0.01883 \\ 0.0002 \\ 0.00001 \end{bmatrix} v[k] + w[k] \\ &= A \bar{x}[k] + B v[k] + w[k] \end{aligned}\]

where $\bar{x}[k]$ is the continuous state vector, $v[k]$ is the automated delivery system control input, and $w[k] \sim \mathcal{N}(0, M)$ is the process noise.

ArchCompStochasticModels.integrator_chain — Method

Integrator-Chain Benchmark

Mathematical Model

For a system of $n$ integrators, the state-space model is given by

\[ x_i[k + 1] = x_i[k] + \sum_{j = i + 1}^n \frac{\tau^{j - i}}{(j - i)!} x_j[k] + \frac{\tau^{n - i + 1}}{(n - i + 1)!} u[k] + w_i[k]\]

where $\tau$ is the sampling time and $w_i[k] \sim \mathcal{N}(0, 0.01)$ is the process noise.

ArchCompStochasticModels.van_der_pol — Method

Stochastic Van der Pol Oscillator Benchmark

First presented in Alessandro, A., Henk, B., Nathalie, C., Joanna, D., Arnd, H., Mahmoud, K., ... & Zuliani, P. (2020). ARCH-COMP20 Category Report: Stochastic Models.

Mathematical Model

\[\begin{aligned} x_1[k + 1] &= x_1[k] + \tau x_2[k] + w[k]\\ x_2[k + 1] &= x_2[k] + \tau ((1 - x_1[k]^2) x_2[k] - x_1[k]) + w[k] \end{aligned}\]